backend/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
drf-nested-routers==0.93.3
psycopg2==2.8.6
PyJWT==2.1.0
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
